requirements.txt
amqp==5.1.1
asgiref==3.6.0
astroid==2.15.8
async-timeout==4.0.2
attrs==22.2.0
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.1
django-qr-code==3.1.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.1
flake8==6.0.0
flake8-django==1.4
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
lazy-object-proxy==1.9.0
Markdown==3.4.3
MarkupSafe==2.1.3
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pluggy==1.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.5
pycodestyle==2.10.0
pydantic==1.10.6
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0.1
redis==4.6.0
requests==2.31.0
segno==1.5.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
